Handle legacy time-on-page metric in query_executor.ex
No behavioral changes
